Revert "Rename index -> fileindex (#83)"
This reverts commit d8c982ea5ddb4cbd09ea1d5bba63daf7c24ed45e.
