+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "tiny-tileserver",
   "description": "Tiny raster+vector tiles and static file server.",
-  "version": "1.3.9",
+  "version": "1.3.8",
   "main": "index.js",
   "engines": {
     "node": ">=10.0.0"
